Implement the stack navigator in app and import it to index
