+++ index.js
@@ -3,8 +3,8 @@
  */
 
 import {AppRegistry} from 'react-native';
+
 import App from './App';
-import Home from './src/containers/Home';
 import {name as appName} from './app.json';
 
-AppRegistry.registerComponent(appName, () => Home);
+AppRegistry.registerComponent(appName, () => App);
